Add new interface function to settle trustlines.
For the need to settle down a trustline, client must be able to access
the relays implementation to find a path with enough capacity to do so.
If a path is found, the library build a transaction for the user.
It was necessary to define a new type `ClosePath` as return object,
cause it vary minimal from the already existing path objects.

The tests set-up a triangle of three users with trustlines in between.
Two cases are tested so far, where once a path could be found and once
it is not possible, cause the opposite direction does not work, due to
the fact that transactions must have positive values.

Related to: #142