Change bump script
Previous script for bumping version was not working as expected.
The command `npm version VERSION --allow-same-version` failed because
the commit was empty.
The problem was that the current version in `package.json` should be
used for bumping but as no changes occured the command failed.
It now prepares the next version via `prepatch` so that changes occur.