Change eth-lightwallet dependency to forked repo
There was an issue with an incorrect peer dependency when using
`yarn install`. It seems there is a new `bitcore-mnemonic` version
which breaks the installation of `eth-lightwallet`. Temporarily using
therefore forked repo where this is fixed. There is also already an open
PR in the `eth-lightwallet` repo with the fix.