add Dockerfile for building trustlines/e2e image
This Dockerfile build an image, which is being used by the end2end repo to run
the end2end tests.

We need a way for the tests to connect to a non-localhost relay server, so make
that configurable via the RELAY_HOSTNAME environment variable.

We use the standard node image instead of trustlines/builder

this results in a much smaller image. saving/restoring took quite long for the
builder based image.

see https://github.com/trustlines-network/clientlib/issues/174