circleci: build and upload e2e image
We use a second workflow for this at the moment, we may need to move this to the
main workflow, depending on how we want to run the end2end tests there.

Also, at the moment we only push the latest tag. At the moment this probably
doesn't matter because we do not have a good way to choose another version when
running the end2end tests in the other repos. So, this is also something we
might change later.