resolve fsevents to 1.2.9 to fix a build issue
When trying to run yarn on node 12 we get an error:
make: *** [Release/obj.target/fse/fsevents.o] Error 1
gyp ERR! build error
gyp ERR! stack Error: `make` failed with exit code: 2

We don’t directly use fsevent, but chokidar does. Watchpack includes chockider and webpack includes watchpack…

Watchpack in version 2 (still in beta) no longer uses chokidar, but we are stuck with 1.6 - this is what webpack has specified. So to fix the issue with building on node v12 I’ve provided a resolve param in the package.json that specifies that we want to use fsevent 1.2.9 and that version builds properly on node v12