Enable ws stream subscribers to handle ws errors
It was not possible to handle ws errors when using the clientlib.
Even though the clientlib handles the reconnection via
[reconnecting-websocket](https://github.com/pladaria/reconnecting-websocket),
one could not know or handle that the ws connection got closed.

It is now possible to pass in options for the reconnection or even
disable it via `reconnectOnError: false`.
It is also possible to detect that a ws connection errored while the
clientlib handles the reconnection.