Use MessagePack to encode extraData
Use MessagePack to encode the extraData of a Transfer. This is so far
only the paymentRequestId. I chose to simply use bytes as the datatype
for now, but that might change later, and we might want to use a custom
type.
The attaching of the decoded paymentRequestId had to be moved, so that
it will be encluded in every endpoint where Transfers are received.
At the moment it can not handle the old extraData and paymentRequestId
at the same time. This is due to the unclear behaviour of what extraData
on the receiver side should contain, if both were included.