Add messageId to prepared payments
The prepared payment will include a messageId as default
use options = { addMessageId: false } to remove it.

Add function confirmPayment that takes a preparedPayment as argument
and confirm the payment as well as should send a message for the
payment.

Deprecate `Payment.confirm`, use `Payment.confirmPayment` instead

Update utils.formatToHexString to return a hex string with an even
number of char, thus accuratly representing bytes.