Update address used in User.ts to getAddress()
This makes the method work with web3providers.
I decided not to remove the `address` method of User, it will raise
an error if called when using a web3provider just as many other
method do (e.g. User.showSeed), and provides for a convenient way
to access the user address when you know you are dealing with a
local wallet.