Make the bump script user friendly
If one ran `yarn bump` the bump script complained that no valid version was supplied, but it continued running and updated the current version and pushed to github. Unfortunately because no version was provided the script wasn’t able to create a git tag. With this change when no argument is provided the script stops executing and asks the user to provide the valid argument.