Fix trustline updates with transfer of "0"
If we are triyng to update a trustline and we pass a transfer value of "0"(string) the clientlib will call the overloaded updateTrustline function of the contract. This causes the meta transaction to fail as there is already an existing trustline and when updating we shouldn't be passing a transfer value.